URIs stored in DFOs are relative, so MyData needs the base path (location) when resuming uploads.  The UploaderRegistrationRequest model has been changed, so that it now links to a StorageBox which includes the base path (location).
New model classes have been added for StorageBox, StorageBoxOptions and
StorageBoxAttributes, which can be found in models/storage.py.
These classes are being used to try out a new way of code structuring -
putting the model classes in their own folder.  This will probably be
applied to all model classes soon, except possibly the data view model
classes which relate to the GUI.

The storage box associated with the UploaderRegistrationRequest approval
is assumed to have attributes for "scp_username" and "scp_hostname".

These changes are related to this pull request (to allow querying of
MyTardis storage box attributes and options via the API):
https://github.com/mytardis/mytardis/pull/445